# api/openapi/jcalendar.yaml
openapi: 3.0.1
info:
  title: Swagger JCalendar API
  description: JCalendar API
  contact: { }
  version: "1.0"
servers:
  - url: /api
paths:
  /users/{id}:
    get:
      summary: Returns user information
      description: Return information about user by ID
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: user ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{email}:
    get:
      summary: Returns user information
      description: Returns user information
      tags:
        - user
      parameters:
        - name: email
          in: path
          description: user email
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    post:
      summary: Adds user information
      description: Adds information about user
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedUser'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{id}:
    get:
      summary: Event information
      description: Returns event information
      tags:
        - events
      parameters:
        - name: id
          in: path
          description: events identificator
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{user_id}:
    get:
      summary: Returns events information for user with user_id
      description: Returns information about enets for specific user
      tags:
        - events
      parameters:
        - name: user_id
          in: path
          description: user identificator
          required: true
          schema:
            type: string
        - name: from
          in: query
          description: begin of interval
          required: true
          schema:
            type: string
        - name: till
          in: query
          description: end of interval
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events:
    post:
      summary: Events information
      description: Adds events information
      tags:
        - events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedEvent'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{id}:
    put:
      summary: Updating status of invite
      description: Updates invite status
      tags:
        - invite
      parameters:
        - name: id
          in: path
          description: invite identificator
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUpdateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedEvent'
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /windows:
    get:
      summary: Returns closets free window for meeting
      description: Returns information about free window for meeting
      tags:
        - windows
      parameters:
        - name: user_ids[]
          in: query
          description: users identificators
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreeWindowResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UserReponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OutputUser'
    UserRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/InputUser'
    OutputUser:
      type: object
      properties:
        ID:
          type: integer
        CreateAt:
          type: string
        UpdateAt:
          type: string
        FirstName:
          type: string
        LastName:
          type: string
        Email:
          type: string
        TimeZoneOffset:
          type: integer
    InputUser:
      type: object
      properties:
        FirstName:
          type: string
        LastName:
          type: string
        Email:
          type: string
        TimeZoneOffset:
          type: integer
    EventRequest:
      properties:
        data:
          $ref: '#/components/schemas/InputEvent'
    EventResponse:
      properties:
        data:
          $ref: '#/components/schemas/OutputEvent'
    EventsResponse:
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OutputEvent'
    InputEvent:
      type: object
      properties:
        From:
          type: string
        Till:
          type: string
        CreatorID:
          type: integer
        Participants:
          type: array
          items:
            type: integer
        Details:
          type: string
        IsPrivate:
          type: boolean
        IsRepeat:
          type: boolean
    OutputEvent:
      type: object
      properties:
        ID:
          type: integer
        CreateAt:
          type: string
        UpdateAt:
          type: string
        From:
          type: string
        Till:
          type: string
        CreatorID:
          type: string
        Details:
          type: string
        IsPrivate:
          type: boolean
    InviteUpdateRequest:
      type: object
      properties:
        IsAccepted:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        Title:
          type: string
        Status:
          type: integer
    FreeWindowResponse:
      properties:
        data:
          $ref: '#/components/schemas/FreeWindow'
    FreeWindow:
      type: object
      properties:
        From:
          type: string
        Till:
          type: string
    CreatedUser:
      type: object
      properties:
        ID:
          type: integer
    CreatedEvent:
      type: object
      properties:
        ID:
          type: integer